requirements.txt
setuptools
requests
pyyaml
pandas==1.4.4
ckanapi
bs4
selenium
selenium-wire
deepl
ftfy